INTSAMPLES-122: Polishing according PR comments
