+++ dsl/si4demo/src/test/java/org/springframework/integration/samples/si4demo/ApplicationTests.java
@@ -1,10 +1,10 @@
-package org.springframework.integration.samples.javaconfig;
+package org.springframework.integration.samples.si4demo;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.integration.samples.javaconfig.annotations.Application;
+import org.springframework.integration.samples.si4demo.annotations.Application;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
