+++ gradle.properties
@@ -1,2 +1,2 @@
 version=2.2.0.BUILD-SNAPSHOT
-springBootVersion=1.1.0.M1
+springBootVersion=1.1.0.M2
